refactor: to get an echarts option through a function call
